API CENTRAL/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
psycopg[binary]==3.1.18
psycopg-pool==3.2.1